chore: refactor didroom java class
