+++ android/app/capacitor.build.gradle
@@ -20,6 +20,7 @@ dependencies {
     implementation project(':capacitor-keyboard')
     implementation project(':capacitor-preferences')
     implementation project(':capacitor-status-bar')
+    implementation project(':capacitor-native-settings')
 
 }
