+++ android/app/src/main/java/com/forkbomb/wallet/MainActivity.java
@@ -1,4 +1,4 @@
-package didroom.forkbomb.eu;
+package com.forkbomb.wallet;
 
 import android.os.Bundle;
