+++ android/app/src/main/java/com/forkbomb/wallet/OldAndroidVersionException.java
@@ -1,4 +1,4 @@
-package didroom.forkbomb.eu;
+package com.forkbomb.wallet;
 
 public class OldAndroidVersionException extends Exception {
     OldAndroidVersionException() {
