+++ android/app/src/main/java/com/forkbomb/wallet/TEE.java
@@ -1,10 +1,11 @@
-package didroom.forkbomb.eu;
+package com.forkbomb.wallet;
 
 import android.os.Build;
 import android.security.keystore.KeyGenParameterSpec;
 import android.security.keystore.KeyProperties;
 import android.util.Pair;
 
+import com.forkbomb.wallet.OldAndroidVersionException;
 import com.getcapacitor.JSArray;
 import com.getcapacitor.JSObject;
 import com.getcapacitor.Plugin;
@@ -14,7 +15,6 @@ import com.getcapacitor.annotation.CapacitorPlugin;
 
 import java.nio.ByteBuffer;
 import java.security.KeyStore;
-import java.security.SecureRandom;
 import java.util.Base64;
 import java.util.Collections;
 import java.util.Enumeration;
@@ -23,7 +23,6 @@ import javax.crypto.Cipher;
 import javax.crypto.KeyGenerator;
 import javax.crypto.SecretKey;
 import javax.crypto.spec.GCMParameterSpec;
-import javax.crypto.spec.IvParameterSpec;
 
 @CapacitorPlugin(name = "TEE")
 public class TEE extends Plugin {
