--- android/local.properties.example
@@ -1,1 +1,0 @@
-sdk.dir=/Users/<HOME_NAME>/Library/Android/sdk
